% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisect_line_segments.R
\name{bisect_line_segments}
\alias{bisect_line_segments}
\title{Bisect Line Segments}
\usage{
bisect_line_segments(rpolyline, projcrs, w, resolution)
}
\arguments{
\item{rpolyline}{depth values along a transect}

\item{projcrs}{coordinate system}

\item{w}{width (m)}

\item{resolution}{resolution of bathymetry data (m)}
}
\value{
cross_section a SpatialLines object containing all the bisectors
}
\description{
Bisects each line segment in a polyline
}
\details{
Calculates the angle of the river centerline and draws a perpedicular bisector passing through it.
bisect_line_segments2 is preferred for the latest version of auto_transects
}
\examples{
bisect_line_segments(rpolyline, projcrs, w, resolution)
}
\keyword{cross}
\keyword{hydraulics,}
\keyword{sections}
\keyword{transects,}
