% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisect_line_segments.R
\name{bisect_line_segments2}
\alias{bisect_line_segments2}
\title{Bisect line segments 2}
\usage{
bisect_line_segments2(rivsplit, w, resolution, mid = FALSE)
}
\arguments{
\item{rivsplit}{river centerline, split into segments}

\item{w}{width to extend the bisectors on either side of the river centerline}

\item{resolution}{resolution of bathymetry data (m)}

\item{mid}{TRUE to use the middle line segment, FALSE to assume the line segment is straight}
}
\description{
Improved version of bisect_line_segments that uses the polyline_splitter from jmt2080ad's github page
}
\details{
When using mid = FALSE, this function implicitly assumes the line segments are not curved. 
The consequence of this is that the bisectors may not be somewhat off center, 
especially when the segments are relatively long. 
If line segments are relatively long, it is better to set mid = TRUE.
}
