% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cross_sections.R
\name{build_cross_sections}
\alias{build_cross_sections}
\title{Build cross sections}
\usage{
build_cross_sections(d, refWSE, wbf)
}
\description{
Stores cross section data (bed elevation, depth, and x coordinate) in a convenient format
}
