% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_A0_pepsi.R
\name{calc_A0_pepsi}
\alias{calc_A0_pepsi}
\title{Calculate A0 for Pepsi Challenge data}
\usage{
calc_A0_pepsi(WSEw, t, A)
}
\arguments{
\item{WSEw}{WSE-w data frame for a reach (not sorted)}

\item{t}{time step}

\item{A}{flow area for reach r at time t}
}
\description{
Calculates A0, the flow area below the observed water surface elevation.
}
\examples{
r <- 1
t <- 365
A0[r,t] <- calc_A0_pepsi(WSEw[[r]], t, A[r,t])
}
