% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prediction_error.R
\name{calc_A0_prediction_error}
\alias{calc_A0_prediction_error}
\title{Calculate submerged area prediction error}
\usage{
calc_A0_prediction_error(pred, truth, A = NULL, type)
}
\arguments{
\item{pred}{minimum bed elevation z0 predictions}

\item{truth}{true minimum bed elevation}

\item{A}{bankfull flow area}

\item{type}{can be "absolute" or "relative"}
}
\description{
Calculates submerged area prediction error, either absolute or as a fraction of bankfull flow area
}
\examples{
A0.l.error.abs <- calc_A0_prediction_error(A0.l, A0.true.ra, type = "absolute")
A0.l.error.rel <- calc_A0_prediction_error(pred = A0.l, truth = A0.true.ra, A = A.true.ra, type = "relative")
}
