% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_A0_variance.R
\name{calc_A0_variance}
\alias{calc_A0_variance}
\title{Calculate variance of the A0 estimate}
\usage{
calc_A0_variance(model, type = "linear", sd_wse = 0.1)
}
\description{
Currently only handles linear models, and also it assumes there is no width measurement error
}
