% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_true_A_WP.R
\name{calc_A_from_WSEw}
\alias{calc_A_from_WSEw}
\title{Calculate true area from WSE-w}
\usage{
calc_A_from_WSEw(WSEw)
}
\arguments{
\item{WSEw}{data frame containing water surface elevation and flow width values}
}
\value{
A flow width
}
\description{
Calculates A from WSE-w relationship only.
}
\details{
Uses trapezoidal rule
This works, much simpler than computing A from cross section data if you already have the WSEw relationship
}
