% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gof.R
\name{calc_gof}
\alias{calc_gof}
\title{Calculate goodness of fit metrics}
\usage{
calc_gof(model, type)
}
\arguments{
\item{model}{a model fit like the result of lm}

\item{type}{type of fit (if sb or nlsb, then the implementation has to change)}
}
\value{
SSE, AIC, BIC, r2, MAE (mean absolute error), mAE (maximum absolute error)
}
\description{
Calculates SSE, AIC, BIC, r2, mean absolute error, and maximum absolute error
for WSE-w fits.
}
