% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean_cross_section.R
\name{calc_mean_cross_section}
\alias{calc_mean_cross_section}
\title{Calculate mean cross section (main)}
\usage{
calc_mean_cross_section(cross_sections, reach_length, section_length)
}
\arguments{
\item{cross_sections}{output from auto_transects}

\item{reach_length}{desired averaging length (m)}

\item{section_length}{distance between cross sections (m)}
}
\value{
cross_section_avg reach average cross sections
}
\description{
Calculate mean cross section (main)
}
