% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean_cross_section.R
\name{calc_mean_cross_section_sub}
\alias{calc_mean_cross_section_sub}
\title{Calculate Mean Cross Section (subroutine)}
\usage{
calc_mean_cross_section_sub(xs.res)
}
\arguments{
\item{xs.res}{cross section geometry that you wish to average, 
resampled to the same length}
}
\description{
A subroutine of calc_mean_cross_section
}
