% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_model_A.R
\name{calc_model_A}
\alias{calc_model_A}
\title{Calculate modeled bankfull flow area}
\usage{
calc_model_A(model, type, WSEw = NULL)
}
\arguments{
\item{model}{fitted model to the WSE-w data}

\item{type}{type of model. Default is linear. Options are linear, sb, sbm, nl, and nlsb}

\item{WSEw}{WSE-w data for the fit. Only required if fit is nonlinear}
}
\value{
A cross-sectional area of flow
}
\description{
Assumes some model, such as linear, nonlinear, or slope break for the channel cross section
}
\examples{
A.l <- calc_modelA(lf[[1]], type = "linear")
A.sb <- calc_modelA(sbm[[1]], type = "sbm")
}
