% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_model_A0.R
\name{calc_model_A0}
\alias{calc_model_A0}
\title{Calculate modeled flow area beneath the lowest observed WSE value}
\usage{
calc_model_A0(model, type, w0 = NULL)
}
\arguments{
\item{model}{fitted model to the WSE-w data}

\item{type}{type of model. Default is linear. Options are linear, sb, sbm, nl, and nlsb}

\item{w0}{minimum observed width value for the fit. Only required if fit is nonlinear}
}
\value{
A0 cross-sectional area of flow beneath lowest observation
}
\description{
Assumes some model, such as linear, nonlinear, or slope break for the channel cross section
}
