% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hydraul_params_model.R
\name{calc_modelA}
\alias{calc_modelA}
\title{Calculate modeled hydraulic parameters}
\usage{
calc_modelA(model, type, WSEw = NULL)
}
\arguments{
\item{model}{fitted model to the WSE-w data}

\item{type}{type of model. Default is linear. Options are linear, sb, sbm, nl, and nlsb}

\item{WSEw}{WSE-w data for the fit. Only required if fit is nonlinear}
}
\value{
A cross-sectional area of flow
}
\description{
Assumes some model, such as linear, nonlinear, or slope break for the channel cross section
}
\details{
Need to update to use the fitted values for WSE, rather than the data values. 
We are calculating parameters for the modeled cross sections, after all.
}
\examples{
True area is 2807
Linear: 1623
SB: 2313
SBm: 2562
NL: 2432
NLSB: 1425
A.l <- calc_modelA(lf[[1]], type = "linear")
A.sb <- calc_modelA(sbm[[1]], type = "sbm")
}
