% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_A0_pepsi.R
\name{calc_overlying_A}
\alias{calc_overlying_A}
\title{Calculates area overlying the lowest observed WSE}
\usage{
calc_overlying_A(WSEw)
}
\arguments{
\item{WSEw}{must be sorted, include only values down to the desired WSE level}
}
\description{
That is, calculates A.prime = A - A0
}
