% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prediction_error.R
\name{calc_prediction_error}
\alias{calc_prediction_error}
\title{Calculate prediction error}
\usage{
calc_prediction_error(pred, truth, A0 = FALSE)
}
\arguments{
\item{A0}{flag if using A0}
}
\description{
Calculates hydraulic parameter prediction error
}
