% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prediction_error.R
\name{calc_s0_prediction_error}
\alias{calc_s0_prediction_error}
\title{Calculate slope prediction error}
\usage{
calc_s0_prediction_error(pred, truth, type)
}
\arguments{
\item{pred}{minimum bed elevation z0 predictions}

\item{truth}{true minimum bed elevation}

\item{type}{can be "absolute" or "relative"}
}
\description{
Calculates slope prediction error, either absolute or _____
}
\examples{
s0.l.error.abs <- calc_A0_prediction_error(s0.l, s0.true.ra, type = "absolute")
}
