% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_WP.R
\name{calc_WP}
\alias{calc_WP}
\title{Calculate Wetted Perimeter}
\usage{
calc_WP(x, b, WSE, method = "wide")
}
\arguments{
\item{x}{distance from horizontal datum}

\item{b}{channel bed elevation}

\item{WSE}{water surface elevation}

\item{method}{"neal" for Jeff Neal's approximation in Neal et al. (2015), "wide" for wide-channel approximation}
}
\description{
Calculates wetted perimeter for a particular water surface elevation, ranging from empty channel to bankfull conditions
}
\details{
Uses a numerical method to approximate the integral of the difference between WSE and bed elevation
The numerical method only counts area where the WSE is higher than the bed elevation
Needs more work to figure out how to implement arbitrary spatial discretization
}
