% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_WSEw.R
\name{calc_WSEw}
\alias{calc_WSEw}
\title{Calculate WSE-w}
\usage{
calc_WSEw(cross_sections, interval = 0.05, dx = 1)
}
\arguments{
\item{cross_sections}{list of x coordinates, bed elevations, and depths for each transect}

\item{interval}{number of points to sample in each cross-section}

\item{dx}{used for computing width with trapezoidal rule}
}
\description{
Calculates WSE-w pairs for each cross section
}
\examples{
calc_WSEw(x, b, interval = 0.05, dx = 1)
}
