% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_WSEw2.R
\name{calc_WSEw2}
\alias{calc_WSEw2}
\title{Calculate WSE-w}
\usage{
calc_WSEw2(cross_sections, interval = 0.05, dx = 1)
}
\arguments{
\item{cross_sections}{list of x coordinates, bed elevations, and depths for each transect}

\item{interval}{number of points to sample in each cross-section. Unlike calc_WSE, interval does not determine the number of data points, which are instead constrained by the SWOT resolution (50m) and the river width}

\item{dx}{used for computing width with trapezoidal rule}
}
\value{
WSEw data frame containing WSE and flow width values
}
\description{
Calculates WSE-w pairs for each cross section.
}
\details{
Starts at w = 100m, the minimum value, and increases until bankfull width in 50 m increments
Not used. Use calc_WSEw, instead.
}
\examples{
xWSEw <- calc_WSEw(cross_sections, interval = 0.05, dx = 1)
}
