% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_WSEw3.R
\name{calc_WSEw3}
\alias{calc_WSEw3}
\title{Calculate WSE-w}
\usage{
calc_WSEw3(cross_sections, dist = "exponential", pars = c(0.9428,
  74.06), dx = 1, n.obs)
}
\arguments{
\item{cross_sections}{list of x coordinates, bed elevations, and depths for each transect}

\item{dist}{type of distribution for SWOT-sampled stage values.}

\item{pars}{vector of distribution parameters}

\item{dx}{used for computing width with trapezoidal rule}

\item{n.obs}{number of SWOT observations to make. The UMRB is observed 1-2 times per 21-day cycle}
}
\value{
WSEw data frame containing WSE and flow width values
}
\description{
Calculates WSE-w pairs for each cross section.
}
\details{
Makes a vector of WSE values from the minimum bed elevation to the bankfull WSE
Computes flow width for each WSE-w value.
Puts the estimated flow widths into the WSEw data frame
Update from old version: Now takes the SWOT sampling into account, instead of using evenly-spaced measurements
Exponential distribution was found to be the best fit to in-situ stage data for pool 21 of the Mississippi River
}
\examples{
xWSEw <- calc_WSEw(cross_sections, dist = "exponential", pars = c(0.9428, 74.06), n.obs = floor(1*365/10))
xWSEw <- calc_WSEw(cross_sections, dist = "gamma", pars = c(9022.2817, 9434.2034), n.obs = floor(1*365/10))
}
