% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_WSEw_from_stage_data.R
\name{calc_WSEw_from_stage_data}
\alias{calc_WSEw_from_stage_data}
\title{Calculate WSEw from stage time series}
\usage{
calc_WSEw_from_stage_data(cross_sections, h, t, tobs, r, sd_w = 0,
  sd_h = 0)
}
\arguments{
\item{cross_sections}{list of cross sections}

\item{h}{gauged water surface elevations}

\item{t}{vector of times corresponding to gauged water surface elevations}

\item{tobs}{vector of times at which to make an observation}

\item{r}{cross section number}
}
\description{
Calculates water surface elevation-width pairs from stage time series to simulate the data available to SWOT
Corrupts the synthetic measurements, as well
}
\details{
Similar to calc_WSEw, but selects WSE values to enter in with based on a time series of gauge measurements.
Also incorporates some of the functionality from observe().
This function is a work in progress.
}
