% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_par.R
\name{clean_by_z0}
\alias{clean_by_z0}
\title{Clean by z0}
\usage{
clean_by_z0(pred, h1)
}
\arguments{
\item{pred}{list of predicted values}

\item{h1}{array of minimum observed heights}
}
\description{
This removes the predictions with z0<h1
}
