% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rmse.R
\name{compute_rmse}
\alias{compute_rmse}
\title{Computes RMSE}
\usage{
compute_rmse(pred_vals, true_vals)
}
\arguments{
\item{pred_vals}{predicted values of bottom elevation, flow area, wetted perimeter, etc. A list with 5 elements.}

\item{true_vals}{true values}
}
\value{
rmse, a data frame showing rmse at each exposure level, averaged across all river reaches
}
\description{
Computes RMSE between the model predictions of hydraulic parameters and their true values
}
\details{
Must be five pred_vals inputs, corresponding to the linear, slope break, multiple slope break, nonlinear, and nlsb methods
}
