% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_xs_format.R
\name{convert_xs_format}
\alias{convert_xs_format}
\title{Convert cross section format}
\usage{
convert_xs_format(x, b, d)
}
\description{
Converts cross section data structure to a list of data frames'
}
