% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polylineSplitter.r
\name{coordBuild}
\alias{coordBuild}
\title{Polyline splitter}
\usage{
coordBuild(spobj)
}
\description{
Splits polylines at equal intervals
}
\details{
Credit to jmt2080ad for these scripts: https://github.com/jmt2080ad/polylineSplitter
Needs plyr and sp packages
}
