% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_separate_channels.R
\name{count_separate_channels}
\alias{count_separate_channels}
\title{Count separate channels}
\usage{
count_separate_channels(t)
}
\arguments{
\item{t}{= transect}
}
\description{
count number of separate channels
}
