% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_separate_channels.R
\name{count_separate_channels_2}
\alias{count_separate_channels_2}
\title{Count separate channels 2}
\usage{
count_separate_channels_2(t, min_width = 1)
}
\arguments{
\item{t}{= transect}
}
\description{
count number of separate channels, can filter by min_width
}
\details{
Can adjust min_width to exclude very narrow channels
}
