% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_widths.R
\name{estimate_widths}
\alias{estimate_widths}
\title{Estimate Widths}
\usage{
estimate_widths(rpolyline, tlength, nseg, resolution = 5)
}
\arguments{
\item{rpolyline}{a polyline crossing the river channel}

\item{tlength}{}

\item{nseg}{}

\item{resolution}{resolution of bathymetry data (m)}
}
\description{
Estimates bankfull widths of transects based on the angle of the transect and the number of pixels in the transect.
A simplification. Hopefully, it is accurate over large enough transects. 
Should probably revisit and make sure this code is reasonably accurate.
}
\examples{
wbf <- estimate_widths(rpolyline, resolution = 5, channel.pix, nseg)
}
\keyword{bankfull}
\keyword{width}
