% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_banks.R
\name{extend_shorter_bank}
\alias{extend_shorter_bank}
\title{Extend shorter bank}
\usage{
extend_shorter_bank(d)
}
\arguments{
\item{d}{list of depths at cross sections}
}
\description{
Assumes the first and last point in the cross section are zero depth (banks)
}
\details{
May want to use a different assumption in the future
}
