% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_xs_wbf.R
\name{extract_xs_wbf}
\alias{extract_xs_wbf}
\title{Extract cross sections and bankfull width}
\usage{
extract_xs_wbf(cross_section, depth)
}
\arguments{
\item{cross_section}{output from bisect_line_segments()}

\item{depth}{bathymetry raster}
}
\description{
Extracts cross sections and bankfull width
}
\details{
This is the same as extract_wbf, except that it also outputs the transects
and has some error catches.
Extracts data underlying the lines in cross_section.
Uses first and last overlapped pixels to estimate bankfull width from 
perpendicular bisectors to the river centerline.
}
