% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_xs_wbf_multi.R
\name{extract_xs_wbf_multi}
\alias{extract_xs_wbf_multi}
\title{Extract cross sections and bankfull width (multiple channel, parallel)}
\usage{
extract_xs_wbf_multi(cross_section, depth, hpc = TRUE, h = 20)
}
\arguments{
\item{cross_section}{output from bisect_line_segments()}

\item{depth}{bathymetry raster}

\item{hpc}{flag for "high performance computing." Always TRUE.}

\item{h}{minimum width (in pixels) to count a channel}
}
\description{
Extracts cross sections and bankfull width in parallel. Sums widths across multichannel transects.
}
\details{
See extract_xs_wbf.
}
