% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bankfull_height.R
\name{find_bankfull_height}
\alias{find_bankfull_height}
\title{Find bankfull height}
\usage{
find_bankfull_height(x, b, delh = 0.01)
}
\arguments{
\item{x}{distance}

\item{b}{bed elevation}

\item{delh}{increment for h}
}
\description{
Finds the height at which overbank flow occurs
}
\details{
Takes (x,b) data at a cross section as input 
Works fairly well, but fails when a cross section has multiple subchannels
}
\examples{
x = cross_sections$x[[7365]]
b = cross_sections$b[[7365]]
hbf <- find_bankfull_height(x,b)
plot(x, b, type = "l")
abline(hbf, 0)
}
