% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plots.R
\name{fit_full_channel}
\alias{fit_full_channel}
\title{Monte Carlo convergence theory}
\usage{
fit_full_channel(savename, plotflag = FALSE)
}
\arguments{
\item{savename}{filename to save the fitted models, goodness of fit metrics, and shape parameters}
}
\description{
Tests whether the Monte Carlo simulations converge to the 
true distribution for the linear case with no measurement error
}
\details{
It is not clear whether this procedure is correct. More details should be worked out.
Also, this code does not work right now, but it shouldn't be too hard to get running. 
Just not a priority.
Fit models to fully-known raech-averaged channel geometry

Fits WSE-w models given fully exposed cross section geometry with no measurement error

fits h-w models to each cross section, assuming full exposure and no measurement error
computes goodness of fit statistics for each model type, across all cross sections, and plots them in histograms
computes shape parameters for each cross section and plots them in a histogram
saves numerical results
}
