% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nlsb.R
\name{fit_nlsb}
\alias{fit_nlsb}
\title{Nonlinear Slope Break Fit}
\usage{
fit_nlsb(WSEw, h = 5)
}
\arguments{
\item{WSEw}{WSEw data (at a given level of exposure)}

\item{h}{minimum distance from either of the WSE-w curve where the breakpoint is constrained to be}
}
\description{
Fits a piecewise nonlinear curve with one breakpoint
There is one breakpoint, it is estimated using strucchange::breakpoints
}
