% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nlsb_try_multi.R
\name{fit_nlsb_try_multi}
\alias{fit_nlsb_try_multi}
\title{Nonlinear Slope Break Fit (multi)}
\usage{
fit_nlsb_try_multi(WSEw, h = 10, maxiter = 100)
}
\arguments{
\item{WSEw}{WSEw data (at a given level of exposure)}

\item{h}{minimum distance from either of the WSE-w curve where the breakpoint is constrained to be}
}
\description{
Fits a piecewise nonlinear curve with one breakpoint
There is one breakpoint, it is estimated using strucchange::breakpoints
}
\details{
Uses multiple guesses for the initial parameters and chooses the best fit
}
