% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_slopebreak.R
\name{fit_slopebreak}
\alias{fit_slopebreak}
\title{Slope Break Fit for WSE-w Relationship}
\usage{
fit_slopebreak(WSEw, mersel = FALSE, thres = 0.015, window = 4,
  multiple_breaks = FALSE, continuity = TRUE, minlen = 5)
}
\arguments{
\item{WSEw}{WSEw data (at a given level of exposure)}

\item{mersel}{flag for Mersel method}

\item{thres}{threshold for Mersel method, must be provided if using Mersel method. 0.015 has been used in the past.}

\item{multiple_breaks}{one slope break or multiple slope breaks}

\item{continuity}{require continuity between piecewise fits?}

\item{minlen}{minimum segment length for breakpoints(). Default is 5.}

\item{thres}{moving average window for Mersel method, must be provided if using Mersel method. Default value is 4.}
}
\description{
Fits any of several types of "slope break" fits to WSE-w data
}
\details{
# if no slope breaks are identified, it just fits a single linear model
Can do the following types of fits
1. Linear with one breakpoint, continuous
2. Linear with one breakpoint, noncontinuous
3. Linear with several breakpoints, continous 
4. Linear with several breakpoints, noncontinuous (returns just the lowest fit)
5. Mersel method: linear with one breakpoint, noncontinous,  screens out "non-optimal" cross sections (returns just the lowest fit)
}
