% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hw.R
\name{generate_hw}
\alias{generate_hw}
\title{Generate h-w}
\usage{
generate_hw(cross_sections, gauge_data, dx = 1, n.obs)
}
\arguments{
\item{cross_sections}{list of x coordinates, bed elevations, and depths for each transect}

\item{dx}{used for computing width with trapezoidal rule}

\item{n.obs}{number of SWOT observations to make. The UMRB is observed 1-2 times per 21-day cycle}

\item{dist}{type of distribution for SWOT-sampled stage values.}

\item{pars}{vector of distribution parameters}
}
\value{
WSEw data frame containing WSE and flow width values
}
\description{
Generates height and width data for each cross section, as it might be sampled by SWOT
}
\details{
Makes a vector of WSE values from the minimum bed elevation to the bankfull WSE
Computes flow width for each WSE-w value.
Puts the estimated flow widths into the WSEw data frame
}
\examples{
xWSEw <- calc_WSEw()
}
