% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_depth_from_lutable.R
\name{get_depth_from_lutable}
\alias{get_depth_from_lutable}
\title{Get Depth from Lookup Table}
\usage{
get_depth_from_lutable(transect, depth)
}
\arguments{
\item{transect}{depth values along a transect}
}
\description{
Extracts depth from a lookup table stored with the bathymetry raster. 
The UMESC bathymetry data are stored using a special format. 
This function decodes that special format and puts the data in units of meters.
}
\examples{
get_depth_from_lutable(transect)
}
\keyword{compression}
\keyword{data}
\keyword{lookup}
\keyword{table,}
