% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_main_channel.R
\name{get_main_channel}
\alias{get_main_channel}
\title{Get Main Channel}
\usage{
get_main_channel(transect)
}
\arguments{
\item{transect}{depth values along a transect}
}
\description{
Uses the transect to get the main channel from a transect potentially covering multiple channels. 
Also provides other useful info, like width to banks.
}
\examples{
get_main_channel(transect)
}
\keyword{cross}
\keyword{hydraulics,}
\keyword{sections}
\keyword{transects,}
