% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maxdiff.R
\name{get_maxdiff}
\alias{get_maxdiff}
\title{Get maxdiff}
\usage{
get_maxdiff(WSE, w)
}
\arguments{
\item{WSE}{WSE}

\item{w}{width}
}
\description{
Gets the maximum difference in the slope between any two points in a series
}
