% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean_cross_section.R
\name{get_start_end_ind}
\alias{get_start_end_ind}
\title{Get start and end indices of reaches}
\usage{
get_start_end_ind(n.xs, reach_length, section_length)
}
\arguments{
\item{n.xs}{output from auto_transects}

\item{reach_length}{desired length of reach (m)}

\item{section_length}{distance between cross sections (m)}
}
\description{
Get start and end indices of reaches
}
