% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_width.R
\name{get_width}
\alias{get_width}
\title{Get Width}
\usage{
get_width(WSE, x, b, delx = 1)
}
\arguments{
\item{WSE}{WSE values with which to enter into the WSE-w function}

\item{delx}{Default is 1.}

\item{f1}{function approximating the WSE-w relationship (like a spline)}

\item{x.max}{}
}
\description{
Numerically estimates the top width
}
\examples{
w <- get_width(WSE, f1, x.max = max(cross.section$x), delx = dx)
}
