% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gof.R
\name{gof_batch}
\alias{gof_batch}
\title{Batch goodness of fit metrics}
\usage{
gof_batch(model, type)
}
\arguments{
\item{lf}{list of model objects}
}
\value{
gof data frame containing goodness of fit measures for each model
}
\description{
Use on a list of models
}
