% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_normal.R
\name{is_normal}
\alias{is_normal}
\title{Test for Gaussianity}
\usage{
is_normal(x)
}
\arguments{
\item{x}{data}

\item{alpha}{significance level}
}
\value{
z 1 if data are Gaussian, 0 otherwise
}
\description{
Uses the Shapiro-Wilks test to test for Gaussianity
}
