% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pepsi_data_from_text.R
\name{load_pepsi_data_from_text}
\alias{load_pepsi_data_from_text}
\title{Load Pepsi data from text file}
\usage{
load_pepsi_data_from_text(data.dir, save.dir, sortWSEw)
}
\arguments{
\item{sortWSE}{boolean, whether or not to sort WSEw from low to high width, 
or leave in the original order from day 1 to nt of the simulation}
}
\description{
Load Pepsi data from text file
}
