% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_shorter_bank.R, R/modify_banks.R
\name{make_zero_bank}
\alias{make_zero_bank}
\title{Make depths at left and right banks zero}
\usage{
make_zero_bank(d)

make_zero_bank(d)
}
\arguments{
\item{d}{list of depths at cross sections}

\item{d}{list of depths at cross sections}
}
\description{
Assumes the first and last point in the cross section are zero depth (banks)

Assumes the first and last point in the cross section are zero depth (banks)
}
\details{
May want to use a different assumption in the future

May want to use a different assumption in the future
}
