% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_one_reach.R
\name{model_one_reach}
\alias{model_one_reach}
\title{Model one reach}
\usage{
model_one_reach(exp_dir, r, expo, sd_wse, sd_w, method = "all",
  full_expo = FALSE, linewidth = 1, ...)
}
\arguments{
\item{full_expo}{boolean for whether or not to exclude observations at less than 100 m width}
}
\value{
A0 predictions from each method
}
\description{
Models one reach, using one or multiple methods
}
