% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_model_A.R
\name{nl_area}
\alias{nl_area}
\title{Calculate area of a nonlinear cross section}
\usage{
nl_area(wbf, WSEbf, z0, a, s)
}
\arguments{
\item{a}{multiplicative parameter for the nonlinear fit}

\item{s}{shape parameter for nonlinear fit (the exponent)}
}
\description{
Calculate area of a nonlinear cross section
}
