% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_model_A.R
\name{nlsb_area4}
\alias{nlsb_area4}
\title{Calculate area of a nonlinear slope break cross section}
\usage{
nlsb_area4(wbf, hbf, z0, a, s, wb)
}
\arguments{
\item{a}{multiplicative parameters for the nonlinear fits}

\item{s}{shape parameters for nonlinear fits (the exponents)}
}
\description{
Calculate area of a nonlinear slope break cross section
}
