% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe.R
\name{observe}
\alias{observe}
\title{Make observations}
\usage{
observe(WSEw, exposure, sd_wse = 0.1, sd_w = 10)
}
\arguments{
\item{exposure}{exposure level as a fraction of bankfull depth}

\item{sd_wse}{standard deviation of expected WSE error}

\item{sd_w}{standard deviation of expected width error}
}
\description{
Sets the exposure level and adds error to the observations.
}
