% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe.R
\name{observe2}
\alias{observe2}
\title{Observe 2}
\usage{
observe2(WSEw, exposure)
}
\description{
Same as observe(), but just performs cutting based on the exposure level.
}
\details{
It does not add error nor censor width measurements under 100 m
}
