% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfit_par.R
\name{observe_par}
\alias{observe_par}
\title{Fit height-width data (parallel)}
\usage{
observe_par(r)
}
\arguments{
\item{r}{cross section number}
}
\value{
list of z0, A, W0, and A0 predictions
}
\description{
Parallelized fitting functions for model fitting with foreach
}
\details{
Saves fitted model (linear, slope break, multiple slope break, nonlinear, or nonlinear slope break) 
and synthetic height-width observations.
}
