% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_predictions_boxplots.R
\name{parameter_predictions_boxplots}
\alias{parameter_predictions_boxplots}
\title{Hydraulic parameter predictions boxplots}
\usage{
parameter_predictions_boxplots(z.l, z.sb, z.nl, z.nlsb, k,
  lumped = FALSE, legend = TRUE, ...)
}
\arguments{
\item{z.l}{linear model parameter predictions or prediction errors}

\item{z.sb}{SB model parameter predictions or prediction errors}

\item{z.nl}{NL model parameter predictions or prediction errors}

\item{z.nlsb}{NLSB model parameter predictions or prediction errors}

\item{k}{exposure level}

\item{lumped}{flag, if true, lump the cross sections together and just compare methods}

\item{legend}{flag for legend}

\item{...}{other parameters as taken by boxplot()}

\item{absolute}{flag for plotting absolute error}
}
\description{
Makes box and whisker plots showing predicted hydraulic parameter error
}
\details{
If lumped = FALSE, plots the results over each cross section separately.
This is the updated version, which can take either predictions or prediction error as input.
}
\examples{
Test case
z.l <- z0.l.error[1:3,,1:2]
z.sb <- z0.sb.error[1:3,,1:2]
z.nl <- z0.nl.error[1:3,,1:2]
z.nlsb <- z0.nlsb.error[1:3,,1:2]
parameter_predictions_boxplots(z.l, z.sb, z.nl, z.nlsb, k = 12, names = bplab)
z.l[,12,]
z.sb[,12,]
z.nl[,12,]
z.nlsb[,12,]

bplab <- c("1","1","1","1","2","2","2","2","3","3","3","3")
lumpedlab <- c("L","SB","NL","NLSB")
par(mfrow = c(2,2))
par(opar)
k <- 16
z0 error
parameter_predictions_boxplots(z0.l.error, z0.sb.error, z0.nl.error, z0.nlsb.error,
                               k = k,
                               lumped = TRUE,
                               main = paste("z0 error at", 100*expo[k], "percent channel exposure"),
                               ylab = "z0 error (m)",
                               legend = FALSE,
                               notch = TRUE,
                               names = lumpedlab)
A0 error
parameter_predictions_boxplots(A0.l.error, A0.sb.error, A0.nl.error, A0.nlsb.error,
                               k = k,
                               lumped = TRUE,
                               main = paste("A0 error at", 100*expo[k], "percent channel exposure"),
                               ylab = "A0 error (sq. m)",
                               legend = FALSE,
                               notch = TRUE,
                               names = lumpedlab)

parameter_predictions_boxplots(z0.l, z0.sb, z0.nl, z0.nlsb, z0.true.ra, k = k, lumped = FALSE,
                               main = paste("z0 error at", 100*expo[k], "percent channel exposure"),
                               ylab = "average z0 error (m)", legend = FALSE, A0 = FALSE, notch = TRUE, names = bplab)
abline(h = z0.true.ra[3])
}
