% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bias.R
\name{plot_bias}
\alias{plot_bias}
\title{Plots bias vs. exposure level}
\usage{
plot_bias(expo, pred_vals, true_vals, na.rm = TRUE, ...)
}
\arguments{
\item{expo}{vector of exposure levels}

\item{pred_vals}{predicted values of bottom elevation, flow area, wetted perimeter, etc. A list with 5 elements.}

\item{true_vals}{true values}

\item{...}{other arguments supported by plot}
}
\value{
bias a data frame showing bias at each exposure level, averaged across all river reaches
}
\description{
Also calculates bias
}
\details{
Must be five pred_vals inputs, corresponding to the linear, slope break, multiple slope break, nonlinear, and nlsb methods
}
\examples{
pred_z0 <- list(z0.l, z0.sb, z0.sbm, z0.nl, z0.nlsb)
plot_bias(expo, pred_z0, z0.true, na.rm = TRUE, 
         main = "z0 bias vs. exposure level, no meas. error", ylab = "Bias (m)", ylim = c(-2,2))
}
