% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_error_vs_exposure.R
\name{plot_brkpt_locs}
\alias{plot_brkpt_locs}
\title{Plot breakpoint locations}
\usage{
plot_brkpt_locs(brk_locs, expo, mean.error, ...)
}
\arguments{
\item{brk_locs}{breakpoint locations}

\item{expo}{vector of exposure levels}

\item{mean.error}{mean error vector returned by plot_error_vs_exposure()}

\item{...}{additional arguments for plot}
}
\description{
Plots breakpoint locations
}
