% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_error_vs_exposure.R
\name{plot_error_vs_exposure}
\alias{plot_error_vs_exposure}
\title{Plot error vs. exposure level}
\usage{
plot_error_vs_exposure(err, ...)
}
\arguments{
\item{err}{z0 error array}

\item{...}{additional arguments to plot()}
}
\value{
mean.error, the mean vector
}
\description{
Plots error vs. exposure level for a given method (L, SB, NL, NLSB)
}
