% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plots.R
\name{plot_gof}
\alias{plot_gof}
\title{Plot goodness-of-fit metrics for a particular class of model}
\usage{
plot_gof(gof, ...)
}
\arguments{
\item{gof}{output of gof_batch}

\item{...}{other arguments as taken by hist}
}
\description{
Used by fit_full_channel
}
