% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plots.R
\name{plot_histograms_of_fitted_params}
\alias{plot_histograms_of_fitted_params}
\title{Histograms of fitted parameters}
\usage{
plot_histograms_of_fitted_params(r, k)
}
\description{
Histograms of fitted parameters
}
\details{
Requires predicted parameter values and true values for comparison
Load predictions without removing the negative A0s (though they are quite similar either way)
}
\examples{
r <- 1
k <- 12 # 60\% exposure
plot_histograms_of_fitted_params(3, 16)
thedir <- "/Volumes/HD3/Cross_Sections/pool_21_ra_10km_nr_3_spacing_5_sampling_even_MC_replicates_500/without_removing_negative_areas"
load(file.path(thedir, "A0_pred_newa0.rda"))
load(file.path(thedir, "z0_pred.rda"))
}
