% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plots.R
\name{plot_histograms_of_param_error}
\alias{plot_histograms_of_param_error}
\title{Histograms of parameter prediction error}
\usage{
plot_histograms_of_param_error(z.l, z.sb, z.nl, z.nlsb, r, k, ...)
}
\description{
Histograms of parameter prediction error
}
\examples{
plot_histograms_of_param_error(A0.l.error, A0.sb.error, A0.nl.error, A0.nlsb.error, r = 2, k = 12, xlim = c(-200,200))
plot_histograms_of_param_error(z0.l.error, z0.sb.error, z0.nl.error, z0.nlsb.error, r = 1, k = 12, xlim = c(-4,1))
}
