% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hw_fits.R
\name{plot_hw_fits}
\alias{plot_hw_fits}
\title{Plot h-w fits for one cross section}
\usage{
plot_hw_fits(WSEw, exposure, sd_wse = 0, sd_w = 0, ...)
}
\arguments{
\item{...}{other parameters as taken by plot()}
}
\value{
predicted hydraulic parameters
}
\description{
Use exposure = 1 for Pepsi Challenge data. Can optionally add measurement error.
}
\examples{
r <- 1
pred <- plot_hw_fits(rWSEw[[r]], exposure = 1, 
                    xlim = c(0, 600), ylim = c(131,148), 
                    main = paste("Reach", r),
                    xlab = "w (m)", ylab = "WSE (m)")
}
