% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hw_fits_2.R
\name{plot_hw_fits_2}
\alias{plot_hw_fits_2}
\title{Plot height-width model 2}
\usage{
plot_hw_fits_2(WSEw, z0, r, type, errorflag = TRUE, exp_dir, k, ...)
}
\arguments{
\item{WSEw}{h-w data}

\item{z0}{z0 predictions, must correpond to the type of model}

\item{r}{reach number}

\item{type}{type of model. Can be "l", "sb", "nl", or "nlsb"}

\item{k}{exposure level index}

\item{...}{additional arguments to plot()}
}
\description{
For making movies of the fitted models with increasing channel exposure
}
\details{
Requires that models have already been fitted and saved as .rds files, 
and that predicted hydraulic parameters (specifically z0) have already been computed.
Must run from the SWOTBATH/home directory.
Currently there is no option for no measurement error.
Currently only has methods for linear model.
}
\examples{
plot_hw_fits_2(rWSEw, z0.l, r = 1, type = "l", ylim=c(138,143), exp_dir = exp_dir, k = 8)
}
