% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ra_pepsi_data.R
\name{plot_ra_pepsi_data}
\alias{plot_ra_pepsi_data}
\title{Plot reach-averaged Pepsi Challenge data}
\usage{
plot_ra_pepsi_data(WSEw.unsorted, A, reach_length = 10000)
}
\description{
May need to edit some of the internal parameters here to get a good plot, e.g. ylim.
}
\examples{
plot_ra_pepsi_data(WSEw.unsorted, A)
}
