% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plots.R
\name{pred_error_hist}
\alias{pred_error_hist}
\title{Histogram of prediction error}
\usage{
pred_error_hist(z, k, ...)
}
\description{
Plots a histogram of the prediction error
}
\examples{
pred_error_hist(z0.sb.error, k = 12, breaks = 100, main = "z0 prediction error, SB, 80\% exposure", xlab = "z0 error (m)")
}
