% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_par.R
\name{pred_nl_par}
\alias{pred_nl_par}
\title{Nonlinear}
\usage{
pred_nl_par(r, WSEw, w1, h1, exclude = FALSE)
}
\arguments{
\item{w1}{minimum observed width, nr by n_exp_level by M array}

\item{h1}{minimum observed height, nr by n_exp_level by M array}
}
\description{
Applies fit_nonlinear in parallel across multiple processors on one machine
}
