% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_predictions.R
\name{process_predictions}
\alias{process_predictions}
\title{Process}
\usage{
process_predictions(pred_lf, pred_sb, pred_nl, pred_nlsb, saveloc)
}
\description{
Processes L, SB, NL, and NLSB predictions to a more convenient format
}
