% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach_avg.R
\name{ra}
\alias{ra}
\title{Reach average general}
\usage{
ra(x, n)
}
\arguments{
\item{x}{quantity, a list}

\item{n}{number of segments making up the reach}
}
\value{
r.out reach averaged quantity
}
\description{
Reach average any quantity
}
\examples{
x <- c(1,2,3,4,4,6)
n <- 3
r <- ra(x,n)
}
