% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach_avg.R
\name{reach_avg}
\alias{reach_avg}
\title{Reach Average}
\usage{
reach_avg(xWSEw, l = 10000, res = 5)
}
\arguments{
\item{xWSEw}{WSEw pairs for each cross section: xWSEw}

\item{l}{= reach length, m. Default is 10 km}

\item{res}{= cross-section spacing, m}
}
\value{
rWSEw Reach averaged WSE and w pairs
}
\description{
Performs reach averaging on the WSE, w data
}
\details{
Tailor made for WSEw data returned by the calc_WSEw function.
}
\examples{
rWSEw <- reach_avg(xWSEw, l = 10000, res = 50)
}
