% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_polyline.R
\name{resample_polyline}
\alias{resample_polyline}
\title{Resample Polyline}
\usage{
resample_polyline(polyline, interval_length = 20,
  add_original_points = TRUE, add_final_point = FALSE)
}
\arguments{
\item{polyline:}{a spatial polyline data frame}

\item{interval_length:}{the length of the segments to split the lines into, in units of the polyline coordinates}

\item{add_original_points:}{whether or not the original points of the polyline should be added to the resulting line
if set FALSE, the resulting line will be shorter}

\item{add_final_point:}{whether or not the final point of the polyline should be added to the resulting line}
}
\value{
df
}
\description{
Functions from agoldev for splitting polylines
}
\details{
#' Source: https://stackoverflow.com/questions/38700246/how-do-i-split-divide-polyline-shapefiles-into-equally-length-smaller-segments
The function splits a polyline into segments of a given length.
}
\examples{
resample_polyline(polyline, interval_length = section_length)
}
\keyword{polyline}
\keyword{resample,}
\keyword{split,}
