% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean_cross_section.R
\name{resample_xs}
\alias{resample_xs}
\title{Resample Cross Section}
\usage{
resample_xs(cross_sections, n)
}
\arguments{
\item{cross_sections}{list of cross section data}

\item{n}{resolution of the resample cross sections}
}
\description{
Uses linear interpolation to resample cross sections with different resolutions
to the same resolution.
}
\examples{
xs.res <- resample_xs(cross_sections, n)
# check that it worked
x <- 100
plot(cross_sections$x[[x]], cross_sections$b[[x]])
lines(xs.res[x,,1], xs.res[x,,2], col = "red")
}
