% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_WSEw.R
\name{rm_null_xs}
\alias{rm_null_xs}
\title{Remove null cross sections}
\usage{
rm_null_xs(xWSEw, return_ind = FALSE)
}
\arguments{
\item{return_ind}{boolean telling the function whether or not to return rm.ind}
}
\description{
Removes null cross sections, those with no measurements (optional)
}
