% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/plot_sample_cross_section.R
\docType{data}
\name{rWSEw}
\alias{rWSEw}
\title{Reach-averaged cross section geometry for Pool 21 of the Upper Mississippi River.}
\format{A list with three entries, one for each reach-averaged cross section. Each list entry contains WSE and w values.
\describe{
  \item{WSE}{water surface elevation, m}
  \item{w}{flow width, m}
}}
\source{
\url{https://www.usgs.gov/centers/umesc}

\url{https://www.usgs.gov/centers/umesc}
}
\usage{
rWSEw

rWSEw
}
\description{
Comes from UMESC bathymetry dataset, processed using WSEw, averaged into three approximate 10 km-long reaches

Comes from UMESC bathymetry dataset, processed using WSEw, averaged into three approximate 10 km-long reaches
}
\keyword{datasets}
