% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin_WSEw.R
\name{sampleWSEw}
\alias{sampleWSEw}
\title{Sample WSEw}
\usage{
sampleWSEw(WSEw, n, first_overpass = 1)
}
\arguments{
\item{WSEw}{WSEw}

\item{n}{number of time steps between successive observations}

\item{first_overpass}{time step of first overpass}
}
\description{
Samples the WSEw data every n timesteps
}
