% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{sb_model_xs}
\alias{sb_model_xs}
\title{Calculate slope break model cross section}
\usage{
sb_model_xs(x, z0, s1, s2, wsb, x0, xf)
}
\arguments{
\item{x}{vector of x coordinates}

\item{z0}{minimum bed elevation}

\item{s1}{slope parameter of the first linear model}

\item{s2}{slope parameter of the second linear model}

\item{wsb}{width at the breakpoint}

\item{x0}{minimum x value, default is 0}

\item{xf}{maximum x value}
}
\value{
b modeled bed elevation
}
\description{
Calculate b(x) for the slope break model
}
\details{
Runs in a loop, could be vectorized for speed
}
