% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_model_A.R
\name{sbm_area}
\alias{sbm_area}
\title{Calculates area of a multi-slope break cross section}
\usage{
sbm_area(w, WSE, z0)
}
\arguments{
\item{w}{width}

\item{WSE}{water surface elevation. The lowest value is z0 and the highest is bankfull WSE.}
}
\description{
Calculates area of a multi-slope break cross section
}
