% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_cross_sections.R
\name{smooth_cross_sections}
\alias{smooth_cross_sections}
\title{Smooth cross sections}
\usage{
smooth_cross_sections(cross_sections, k)
}
\description{
Smooths bed elevation and depth data using a moving average
}
