% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_objects.R
\name{subset_cross_sections}
\alias{subset_cross_sections}
\title{Subset cross sections}
\usage{
subset_cross_sections(cross_sections, ind)
}
\arguments{
\item{cross_sections}{from build_cross_sections}

\item{ind}{index of cross sections to keep}
}
\description{
Subsets the variable cross_sections using an index, ind
}
