% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_objects.R
\name{subset_rWSEw}
\alias{subset_rWSEw}
\title{Subset cross sections}
\usage{
subset_rWSEw(rWSEw, ind)
}
\arguments{
\item{ind}{index of cross sections to keep}

\item{cross_sections}{from build_cross_sections}
}
\description{
Subsets the variable cross_sections using an index, ind
}
