% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_channels.R
\name{test}
\alias{test}
\title{Test}
\usage{
test(x, y, w, span, plotflag = TRUE)
}
\arguments{
\item{w}{tuning parameter}

\item{span}{tuning parameter}
}
\description{
Called Test because it's supposed to be a way to find good tuning parameters using trial and error
}
