% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_linear.R
\name{test_linear}
\alias{test_linear}
\title{Test Linear}
\usage{
test_linear(WSE, w, thres = 0.015)
}
\description{
Tests whether the linear method of Mersel et al. (2013) is appropriate for the WSE-w measurements
}
