% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_slopebreak.R
\name{test_slope_break}
\alias{test_slope_break}
\title{Slope Break Test}
\usage{
test_slope_break(WSE, w, thres = 0.015, window = 4, m = FALSE)
}
\arguments{
\item{thres}{threshold for maxdiff in slope for the WSE-w relationship to be considered linear}

\item{m}{flag for multiple slope breaks}
}
\value{
If it is an optimal location, returns the slope break location. Otherwise returns NULL.
}
\description{
Tests whether there is an "optimal location" for the slope break method, as described by Mersel et al. (2013)
}
