% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin_WSEw.R
\name{thinWSEw}
\alias{thinWSEw}
\title{Thin WSEw}
\usage{
thinWSEw(WSEw, n)
}
\arguments{
\item{WSEw}{WSEw}

\item{n}{number of data points in each of the new WSEw[[r]] data frames}
}
\description{
Thins out the WSEw data if they are too dense
}
