% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_banks.R
\name{truncate_taller_bank}
\alias{truncate_taller_bank}
\title{Truncate taller bank}
\usage{
truncate_taller_bank(d)
}
\arguments{
\item{d}{list of depths at cross sections}
}
\description{
Assumes the first and last point in the cross section are zero depth (banks)
}
\details{
May want to use a different assumption in the future
}
