% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_WSEw.R
\name{update_WSEw}
\alias{update_WSEw}
\title{Updates the WSEw package}
\usage{
update_WSEw()
}
\description{
Run from the /Users/.../Cross_Sections directory with devtools and roxygen2 packages loaded
}
\details{
Sometimes you need to restart the R session in RStudio for the documentation to load properly
}
