% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_transects.R
\name{auto_transects}
\alias{auto_transects}
\title{Auto Transects}
\usage{
auto_transects(section_length, riv, depth, refWSE, savename,
  halfwidth = 1000, k = 5, makeplot = FALSE)
}
\arguments{
\item{section_length}{spatial discretization for dividing the river into cross sections (m)}

\item{riv}{polyline file for the main river channel}

\item{depth}{raster of depth values (m)}

\item{refWSE}{Reference WSE for the measured bathymetry data (m)}

\item{savename}{where to save the outputs}

\item{halfwidth}{guess for how wide the channel is. Default is 1000 m.}

\item{k}{smoothing width}

\item{makeplot}{whether or not to make a plot. Default is FALSE.}
}
\description{
Builds cross sections from bathymetry data
}
\examples{
auto_transects(section_length, savename = "p21.rda")
}
\keyword{cross}
\keyword{hydraulics,}
\keyword{sections}
\keyword{transects,}
