% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_transects.R
\name{auto_transects}
\alias{auto_transects}
\title{Auto Transects}
\usage{
auto_transects(section_length, riv, depth, halfwidth, saveflag = FALSE,
  output_all = FALSE, savedir = NULL, main_only = TRUE)
}
\description{
Draws transects, finds main channel, and gets depth values
}
