% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cross_sections.R
\name{build_cross_sections}
\alias{build_cross_sections}
\title{Build cross sections}
\usage{
build_cross_sections(d, refWSE, wbf, option = "none")
}
\arguments{
\item{option}{Use "zero" to make the left and right banks have depths zero
Use "extend" to extend the shorter bank height to match the taller bank height
Use "truncate" to truncate the taller bank height to match the shorter bank height.}
}
\description{
Stores cross section data (bed elevation, depth, and x coordinate) in a convenient format
}
