% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_A0_variance.R
\name{calc_A0_variance}
\alias{calc_A0_variance}
\title{Calculate A0 variance}
\usage{
calc_A0_variance(m, sd_h = 0.1)
}
\arguments{
\item{m}{linear model}

\item{sd_h}{standard deviation of WSE measurement error (m)}
}
\description{
Calculates A0 variance for the linear case
}
\details{
Assumes no width measurement error
Assumes that WSE measurement errors are independent of WSE
}
