% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gof.R
\name{calc_gof}
\alias{calc_gof}
\title{Calculate goodness of fit metrics}
\usage{
calc_gof(model, type)
}
\arguments{
\item{model}{a model fit like the result of lm}

\item{type}{type of fit (if sb or nlsb, then the implementation has to change)}
}
\value{
gof data frame containing SSE, spectral radius, AIC, BIC, r2, MAE (mean absolute error), and mAE (maximum absolute error)
}
\description{
Calculates SSE, AIC, BIC, r2, mean absolute error, and maximum absolute error
for WSE-w fits.
}
\details{
Calculates goodness-of-fit metrics for a model with respect to the fitted values. 
R2 is not defined for nonlinear models, but it can be computed for a piecewise linear model/
Piecewise AIC and BIC are computed by summing the AIC, BIC values of the original models. I am like 90% sure this is OK.
Spectral radius of the covariance matrix is a measurement of parameter estimation uncertainty. 
However, I am not sure how to calculate it for the piecewise fits, so it just returns NA for now.
}
