% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean_cross_section.R
\name{calc_mean_cross_section_sub}
\alias{calc_mean_cross_section_sub}
\title{Calculate Mean Cross Section (subroutine)}
\usage{
calc_mean_cross_section_sub(xs.res)
}
\arguments{
\item{xs.res}{cross section geometry that you wish to average, 
resampled to the same length}
}
\description{
Uses linear interpolation to calculate the average cross section from 
among individual cross sections with different lengths
}
\examples{
xs.avg <- calc_mean_cross_section(xs.res, n = 100)
calc_mean_cross_section(xs.res[1:100,,])
}
