% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_WSEw.R
\name{calc_WSEw}
\alias{calc_WSEw}
\title{Calculate WSE-w}
\usage{
calc_WSEw(cross_sections, interval = 0.05, dx = 1)
}
\arguments{
\item{cross_sections}{list of x coordinates, bed elevations, and depths for each transect}

\item{interval}{number of points to sample in each cross-section. Unlike calc_WSE, interval does not determine the number of data points, which are instead constrained by the SWOT resolution (50m) and the river width}

\item{dx}{used for computing width with trapezoidal rule}
}
\value{
WSEw data frame containing WSE and flow width values
}
\description{
Calculates WSE-w pairs for each cross section
}
\details{
Makes a vector of WSE values from the minimum bed elevation to the bankfull WSE
Computes flow width for each WSE-w value.
Puts the estimated flow widths into the WSEw data frame
}
\examples{
xWSEw <- calc_WSEw(cross_sections, interval = 0.05, dx = 1)
}
