% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterize_channel.R
\name{characterize_channel}
\alias{characterize_channel}
\title{Characterize channel}
\usage{
characterize_channel(cross_sections, xWSEw, savename = NULL,
  plotflag = FALSE, section_length = 5, na.rm = FALSE)
}
\arguments{
\item{cross_sections}{cross section geometry}

\item{xWSEw}{WSE-w data}

\item{savename}{filename to save the fitted models, goodness of fit metrics, and shape parameters}
}
\description{
Characterizes channel geometry with shape parameter and bankfull parameters
}
\details{
Reports the same metrics as Grimaldi et al. (2018)
Drops the first d-w data point to perform fit and avoid an issue with log(0)
}
