% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_channels.R
\name{count_channels}
\alias{count_channels}
\title{Count channels}
\usage{
count_channels(x, d)
}
\arguments{
\item{x}{x coordinate}

\item{b}{bed elevation}
}
\description{
Returns the number of channels contained within a (possibly multi-channel) cross section
}
\details{
Finds the local minima in the channel cross section
Uses smoothing techniques described at https://rpubs.com/mengxu/peak_detection
Accuracy depends on choice of tuning parameters
Could calibrate to a particular dataset
}
\examples{
k <- 1000
x <- cross_sections$x[[k]]
b <- cross_sections$b[[k]]
d <- cross_sections$d[[k]]
plot(x,b,type="l")
}
