% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_slopebreak.R
\name{fit_slopebreak}
\alias{fit_slopebreak}
\title{Slope Break Fit for WSE-w Relationship}
\usage{
fit_slopebreak(WSEw, mersel = FALSE, thres = NULL,
  multiple_breaks = FALSE, continuity = TRUE)
}
\arguments{
\item{WSEw}{WSEw data (at a given level of exposure)}

\item{mersel}{flag for Mersel method}

\item{thres}{threshold for Mersel method, must be provided if using Mersel method. 0.015 has been used in the past.}

\item{multiple_breaks}{one slope break or multiple slope breaks}

\item{continuity}{require continuity between piecewise fits?}
}
\description{
Optionally uses Mersel's original method, which screens out "non-optimal" cross sections
}
