% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pepsi_data_from_text.R
\name{load_pepsi_data_from_text}
\alias{load_pepsi_data_from_text}
\title{Load Pepsi data from text file}
\usage{
load_pepsi_data_from_text(data.dir, save.dir)
}
\description{
Load Pepsi data from text file
}
\details{
Saves two versions of the h-w data: one is sorted from low to high width, 
and the other is left in the original simulated order of day 1 to nt.
}
