% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plots.R
\name{parameter_predictions_scatterplots}
\alias{parameter_predictions_scatterplots}
\title{Parameter predictions scatterplots (without error)}
\usage{
parameter_predictions_scatterplots(z0.l, z0.sb, z0.nl, z0.nlsb)
}
\description{
Plot predicted z0 and A0 from each model for each reach-average cross section
Assumes there is M = 1 replicate. This should be the no error case.
This is an earlier iteration of Figure 5
}
