% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_nl_par.R
\name{pred_nl_par}
\alias{pred_nl_par}
\title{Nonlinear}
\usage{
pred_nl_par(r, WSEw, w1, h1, exclude = FALSE)
}
\arguments{
\item{r}{cross section number}

\item{w1}{minimum observed width, nr by n_exp_level by M array}

\item{h1}{minimum observed height, nr by n_exp_level by M array}

\item{exclude}{boolean that determines whether or not to exclude predictions from physically unrealistic model fits}
}
\value{
list of z0, A, W0, and A0 predictions
}
\description{
Applies fit_nonlinear in parallel across multiple processors on one machine
}
\details{
Loads fitted model (linear, slope break, multiple slope break, nonlinear, or nonlinear slope break)
and uses it to predict hydraulic parameters
error.flag values: 0 = no error, 1 = negative slope, 2 = negative concavity
There is a problem when s is negative because the predictions at w = 0 are -Inf
}
